Fixed a warning in server's console.log about missing log4j.jar and log4j-core.jar. if left alone, it would not have caused a runtime error, but I fixed it none the less.
I also updated the daytrader-ee6 project and its pom to create a docker
image for the server and its server configuration and its application. I
comented out that change in the pom because i did not want it to run on
each time you package the ear; at least not for now.